Consolidated all the AddressRange/MemoryRange/BulkMemoryRange types. Will need to split out the one used by the emulation for speed. Can now add ranges for the memory dump page. Fixed most of the enumerators to use uint for address and ulong for length.

+++ Jammy.Core/Memory/MemoryMapper.cs
@@ -138,7 +138,7 @@ namespace Jammy.Core.Memory
 		{
 			//todo: expensive!
 			byte[] find = GetEnumerable(0).ToArray();
-			for (int i = 0; i < Length - bytes.Length; i++)
+			for (int i = 0; i < (int)Length - bytes.Length; i++)
 			{
 				if (bytes.SequenceEqual(find.Skip(i).Take(bytes.Length)))
 					return (uint)i;
@@ -147,18 +147,18 @@ namespace Jammy.Core.Memory
 			return 0;
 		}
 
-		public IEnumerable<byte> GetEnumerable(int start, long length)
+		public IEnumerable<byte> GetEnumerable(uint start, ulong length)
 		{
-			for (long i = start; i < Math.Min(start + length, memoryRange.Length); i++)
+			for (ulong i = start; i < Math.Min(start + length, memoryRange.Length); i++)
 				if (memoryManager.DebugMappedDevice[(uint)i] is IUnmappedMemory)
 					yield return 0;
 				else
 					yield return UnsafeRead8((uint)i);
 		}
 
-		public IEnumerable<byte> GetEnumerable(int start)
+		public IEnumerable<byte> GetEnumerable(uint start)
 		{
-			for (long i = start; i < memoryRange.Length; i++)
+			for (ulong i = start; i < memoryRange.Length; i++)
 			{
 				if (memoryManager.DebugMappedDevice[(uint)i] is IUnmappedMemory)
 					yield return 0;
@@ -167,25 +167,25 @@ namespace Jammy.Core.Memory
 			}
 		}
 
-		public IEnumerable<uint> AsULong(int start)
+		public IEnumerable<uint> AsULong(uint start)
 		{
-			for (long i = start; i < memoryRange.Length; i += 4)
+			for (ulong i = start; i < memoryRange.Length; i += 4)
 				if (memoryManager.DebugMappedDevice[(uint)i] is IUnmappedMemory)
 					yield return 0;
 				else
 					yield return UnsafeRead32((uint)i);
 		}
 
-		public IEnumerable<ushort> AsUWord(int start)
+		public IEnumerable<ushort> AsUWord(uint start)
 		{
-			for (long i = start; i < memoryRange.Length; i += 2)
+			for (ulong i = start; i < memoryRange.Length; i += 2)
 				if (memoryManager.DebugMappedDevice[(uint)i] is IUnmappedMemory)
 					yield return 0;
 				else
 					yield return UnsafeRead16((uint)i);
 		}
 
-		public int Length => (int)memoryRange.Length;
+		public ulong Length => memoryRange.Length;
 
 		public List<BulkMemoryRange> GetBulkRanges()
 		{
